ported over blog post page, done with porting
