+++ src/app/shared-metadata.tsx
@@ -1,0 +1,4 @@
+export const openGraph = {
+  title: "Duke Tran | Portfolio",
+  description: "Learn more about Duke Tran and his projects.",
+};
